ConvolutionalClosure

Documentation for ConvolutionalClosure.

ConvolutionalClosure.Burgers

ConvolutionalClosure.Convection

ConvolutionalClosure.Diffusion

ConvolutionalClosure.FourierLayer

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.Schrodinger

ConvolutionalClosure.apply_filter

ConvolutionalClosure.apply_stencils

ConvolutionalClosure.apply_stencils_nonsquare

ConvolutionalClosure.apply_stencils_transpose

ConvolutionalClosure.circulant

ConvolutionalClosure.constant_interpolator

ConvolutionalClosure.convolutional_closure

ConvolutionalClosure.convolutional_matrix_closure

ConvolutionalClosure.create_data

ConvolutionalClosure.derivative_loss

ConvolutionalClosure.eqname

ConvolutionalClosure.extend1D

ConvolutionalClosure.extend2D

ConvolutionalClosure.fourier_closure

ConvolutionalClosure.gaussian

ConvolutionalClosure.gaussian

ConvolutionalClosure.glorot_uniform_Float64

ConvolutionalClosure.linear_interpolator

ConvolutionalClosure.plotfield

ConvolutionalClosure.plotmat

ConvolutionalClosure.plotsol

ConvolutionalClosure.relerr

ConvolutionalClosure.rk4

ConvolutionalClosure.solve_equation

ConvolutionalClosure.top_hat

ConvolutionalClosure.trajectory_loss

ConvolutionalClosure.Burgers — Type

(::Burgers)(u, p, t)

Compute right hand side of Burgers equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Convection — Type

(::Convection)(u, p, t)

Compute right hand side of convection equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Diffusion — Type

(::Diffusion)(u, p, t)

Compute right hand side of diffusion equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.FourierLayer — Type

FourierLayer(n, kmax; σ = identity, init_weight = glorot_uniform_Float64)

Fourier layer operating on uniformly discretized functions with n pointwise output values. Inputs and outputs are both arrays of size (n, nx, nsample), where nx is any number of discrete uniform spatial points that is higher than the cut-off frequency kmax, and the nsample samples are treated independently.

ConvolutionalClosure.KortewegDeVries — Type

KortewegDeVries(l, N = 2)

Korteweg-De Vries equation with domain length l and order N.

ConvolutionalClosure.KortewegDeVries — Method

Compute right hand side of Korteweg-De Vries equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.KuramotoSivashinsky — Type

KuramotoSivashinsky(l, N = 2)

Kuramoto-Sivashinsky equation with domain length l and order N.

ConvolutionalClosure.KuramotoSivashinsky — Method

Compute right hand side of Kuramoto-Sivashinsky equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Schrodinger — Type

(::Schrodinger)(u, p, t)

Compute right hand side of Schrodinger equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.apply_filter — Method

apply_filter(W, u)

Apply filter to a collection of solutions.

ConvolutionalClosure.apply_stencils — Method

apply_stencils(u, p)

Apply the periodically extended multi-diagonal matrix circdiag(-r => p[:, 1], ..., r => p[:, end]) to u for the M stencils p of size (M, 2r+1).

ConvolutionalClosure.apply_stencils_nonsquare — Method

apply_stencils_nonsquare(u, p)

For the M stencils p of size (M, 2r+1).

ConvolutionalClosure.apply_stencils_transpose — Method

apply_stencils_transpose(u, p)

Apply the transpose of the periodically extended multi-diagonal matrix circdiag(-r => p[:, 1], ..., r => p[:, end]) to u for the M stencils p of size (M, 2r+1).

ConvolutionalClosure.circulant — Method

circulant(n, inds, stencil)

Create circulant SparseMatrixCSC.

ConvolutionalClosure.constant_interpolator — Method

constant_interpolator(x, y)

Create matrix for interpolating from grid $x \in \mathbb{R}^N$ to $y \in \mathbb{R}^M$. This function uses a zeroth order interpolation, i.e. the function is piece-wise constant.

ConvolutionalClosure.convolutional_closure — Method

convolutional_closure(r, c, a, b; rng = Random.default_rng, channel_augmenter = identity)

Create closure function with

Args:

r: Kernel radii (nlayer)

c: Number of channels (nlayer + 1) (first should be equal to number of output channels in channel_augmenter)

a: Activation functions (nlayer)

b: Bias indicators (nlayer)

Kwargs:

rng: Random number generator

channelaugmenter: A function `u -> hcat(channel1(u), ..., channel_n(u))`

Return (p, closure), where p are the initial parameters and closure(u, p, t) computes the closure term.

ConvolutionalClosure.convolutional_matrix_closure — Method

convolutional_matrix_closure(r, c, a, b; rng = Random.default_rng, channel_augmenter = identity)

Create closure function with

Args:

r: Kernel radii (nlayer)

c: Number of channels (nlayer + 1) (first should be equal to number of output channels in channel_augmenter)

a: Activation functions (nlayer)

b: Bias indicators (nlayer)

Kwargs:

rng: Random number generator

channelaugmenter: A function `u -> hcat(channel1(u), ..., channel_n(u)`

Return (p, closure), where p are the initial parameters and closure(u, p, t) computes the closure term.

ConvolutionalClosure.create_data — Method

create_data(x, K, nsample; decay = Returns(1))

Create nsample random signals on x with maximum frequency K and frequency decay decay(k).

ConvolutionalClosure.derivative_loss — Method

derivative_loss(f, p, dudt, u; nuse = size(u, 2), λ = 0)

Compute derivative-fitting loss. Chooses a random subset (nuse) of the data samples at each evaluation. Note that both u and dudt are of size (sample_size..., nsample).

ConvolutionalClosure.eqname — Function

eqname(equation)

Get name of equation.

ConvolutionalClosure.extend1D — Method

extend1D(u, r)

Extend the vector u periodically with an overlap of r.

ConvolutionalClosure.extend2D — Function

extend2D(u, rx, ry = rx)

Extend the matrix u periodically with an overlap of (rx, ry).

ConvolutionalClosure.fourier_closure — Method

fourier_closure(n, kmax; channel_augmenter = identity, σ = gelu, [rng])

Fourier neural operator closure with n latent channels, wavenumber truncation at kmax, and actiavation function σ.

In addition to an intial and final point-wise lift and compression, this model is comprised of four inner Fourier layers.

For each Fourier layer, the output is σ(y + z), where

y is a spatial point-wise linear transform of the input

z is a spectral mode-wise linear transform of the truncated spectrum of the input

Note: Spatial inputs can have any uniform discretization, but the number of discretization points must be at least kmax.

ConvolutionalClosure.gaussian — Method

gaussian(Δx, Δy, x, y)

Gaussian filter weight at point (x, y) for filter width (Δx, Δy).

ConvolutionalClosure.gaussian — Method

gaussian(Δ, x)

Gaussian filter weight at point x for filter width Δ.

ConvolutionalClosure.glorot_uniform_Float64 — Method

Generate random weights as in Lux.glorot_uniform, but with Float64. https://github.com/avik-pal/Lux.jl/blob/51bbf8dc489155c53f5f034b636848bdaabfc55d/src/utils.jl#L45-L48

ConvolutionalClosure.linear_interpolator — Method

linear_interpolator(x, y)

Create matrix for interpolating from grid $x \in \mathbb{R}^N$ to $y \in \mathbb{R}^M$.

ConvolutionalClosure.plotfield — Method

plotfield(x, y, u; kwargs...)

Plot solution u.

ConvolutionalClosure.plotmat — Method

plotmat(A; kwargs...)

Plot matrix.

ConvolutionalClosure.plotsol — Method

plotsol(x, t, u; kwargs...)

Plot solution u.

ConvolutionalClosure.relerr — Method

relerr(u, v)

Average relative errors along first dimension.

ConvolutionalClosure.rk4 — Method

rk4(f, p, u₀, t, Δt)

Solve ODE $du/dt = f(u, p, t)$ with RK4, where $p$ are parameters.

ConvolutionalClosure.solve_equation — Method

solve_equation(f, u₀, p, t; kwargs...)

Solve equation from t[1] to t[end]. The ODE right hand side f should work on one vector with size (N, 1) and with multiple vectors in parallel in in a matrix of size (N, nvector). The initial conditions u₀ are of size N × n_IC. The solution is saved at t of size n_t.

Return an ODESolution acting as an array of size N × n_IC × n_t.

ConvolutionalClosure.top_hat — Method

top_hat(Δ, x)

Top-hat filter weight at point x for filter width Δ.

ConvolutionalClosure.trajectory_loss — Method

trajectory_loss(
    f,
    p,
    u,
    t;
    nsolution = size(u, ndims(u) - 1),
    ntime = length(t),
    λ = 0,
    kwargs...,
)

Compute trajectory-fitting loss. Chooses a random subset of the solutions and time points at each evaluation. Note that u is of size (nx, nsolution, ntime)