ConvolutionalClosure

Documentation for ConvolutionalClosure.

ConvolutionalClosure.Burgers

ConvolutionalClosure.Convection

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.apply_filter

ConvolutionalClosure.gaussian

ConvolutionalClosure.loss_derivative_fit

ConvolutionalClosure.loss_embedded

ConvolutionalClosure.relerr

ConvolutionalClosure.top_hat

ConvolutionalClosure.Burgers — Type

(::Burgers)(u, p, t)

Compute right hand side of Burgers equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Convection — Type

(::Convection)(u, p, t)

Compute right hand side of convection equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.KortewegDeVries — Type

KortewegDeVries(l, N = 2)

Korteweg-De Vries equation with domain length l and order N.

ConvolutionalClosure.KortewegDeVries — Method

Compute right hand side of Korteweg-De Vries equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.KuramotoSivashinsky — Type

KuramotoSivashinsky(l, N = 2)

Kuramoto-Sivashinsky equation with domain length l and order N.

ConvolutionalClosure.KuramotoSivashinsky — Method

Compute right hand side of Kuramoto-Sivashinsky equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.apply_filter — Method

Apply filter to a collection of solutions.

ConvolutionalClosure.gaussian — Method

gaussian(Δ, x)

Gaussian filter weight at point x for filter width Δ.

ConvolutionalClosure.loss_derivative_fit — Method

loss_derivative_fit(f, p, dudt, u, λ)

Compute derivative-fitting loss.

ConvolutionalClosure.loss_embedded — Method

loss_embedded(s, p, u, t, λ; kwargs...)

Compute trajectory-fitting loss.

ConvolutionalClosure.relerr — Method

relerr(u, v, t)

Average relative errors.

ConvolutionalClosure.top_hat — Method

top_hat(Δ, x)

Top-hat filter weight at point x for filter width Δ.