ConvolutionalClosure

Documentation for ConvolutionalClosure.

ConvolutionalClosure.Burgers

ConvolutionalClosure.Convection

ConvolutionalClosure.Diffusion

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KortewegDeVries

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.KuramotoSivashinsky

ConvolutionalClosure.apply_filter

ConvolutionalClosure.convolutional_closure

ConvolutionalClosure.create_data

ConvolutionalClosure.derivative_loss

ConvolutionalClosure.eqname

ConvolutionalClosure.gaussian

ConvolutionalClosure.glorot_uniform_Float64

ConvolutionalClosure.plotmat

ConvolutionalClosure.plotsol

ConvolutionalClosure.relerr

ConvolutionalClosure.rk4

ConvolutionalClosure.solve_equation

ConvolutionalClosure.top_hat

ConvolutionalClosure.trajectory_loss

ConvolutionalClosure.Burgers — Type

(::Burgers)(u, p, t)

Compute right hand side of Burgers equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Convection — Type

(::Convection)(u, p, t)

Compute right hand side of convection equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.Diffusion — Type

(::Diffusion)(u, p, t)

Compute right hand side of diffusion equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.KortewegDeVries — Type

KortewegDeVries(l, N = 2)

Korteweg-De Vries equation with domain length l and order N.

ConvolutionalClosure.KortewegDeVries — Method

Compute right hand side of Korteweg-De Vries equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.KuramotoSivashinsky — Type

KuramotoSivashinsky(l, N = 2)

Kuramoto-Sivashinsky equation with domain length l and order N.

ConvolutionalClosure.KuramotoSivashinsky — Method

Compute right hand side of Kuramoto-Sivashinsky equation. This works for both vector and matrix u (one or many solutions).

ConvolutionalClosure.apply_filter — Method

Apply filter to a collection of solutions.

ConvolutionalClosure.convolutional_closure — Method

convolutional_closure(r, c, a; rng = Random.default_rng, channel_augmenter = identity)

Create closure function with

Args:

r: Kernel radii (nlayer)

c: Number of channels (nlayer + 1) (first should be equal to number of output channels in channel_augmenter)

a: Activation functions (nlayer)

b: Bias indicators (nlayer)

Kwargs:

rng: Random number generator

channelaugmenter: A function `u -> hcat(channel1(u), ..., channel_n(u)`

Return (p, closure), where p are the initial parameters and closure(u, p, t) computes the closure term.

ConvolutionalClosure.create_data — Method

create_data(x, K, nsample; decay = Returns(1))

Create nsample random signals on x with maximum frequency K and frequency decay decay(k).

ConvolutionalClosure.derivative_loss — Method

Compute derivative-fitting loss. Chooses a random subset (nuse) of the data samples at each evaluation. Note that both u and dudt are of size (nx, nsample).

ConvolutionalClosure.eqname — Function

eqname(equation)

Get name of equation.

ConvolutionalClosure.gaussian — Method

gaussian(Δ, x)

Gaussian filter weight at point x for filter width Δ.

ConvolutionalClosure.glorot_uniform_Float64 — Method

Generate random weights as in Lux.glorot_uniform, but with Float64. https://github.com/avik-pal/Lux.jl/blob/51bbf8dc489155c53f5f034b636848bdaabfc55d/src/utils.jl#L45-L48

ConvolutionalClosure.plotmat — Method

plotmat(A; kwargs...)

Plot matrix.

ConvolutionalClosure.plotsol — Method

plotsol(x, t, u; kwargs...)

Plot solution u.

ConvolutionalClosure.relerr — Method

relerr(u, v)

Average relative errors along first dimension.

ConvolutionalClosure.rk4 — Method

rk4(f, p, u₀, t, Δt)

Solve ODE $du/dt = f(u, p, t)$ with RK4, where $p$ are parameters.

ConvolutionalClosure.solve_equation — Method

solve_equation(f, u₀, p, t; kwargs...)

Solve equation from t[1] to t[end]. The ODE right hand side f should work on one vector with size (N, 1) and with multiple vectors in parallel in in a matrix of size (N, nvector). The initial conditions u₀ are of size N × n_IC. The solution is saved at t of size n_t.

Return an ODESolution acting as an array of size N × n_IC × n_t.

ConvolutionalClosure.top_hat — Method

top_hat(Δ, x)

Top-hat filter weight at point x for filter width Δ.

ConvolutionalClosure.trajectory_loss — Method

Compute trajectory-fitting loss. Chooses a random subset of the solutions and time points at each evaluation. Note that u is of size (nx, nsolution, ntime)